pyproject.toml
[project]
name = "yao"
version = "0.1.0"
description = "YaO — An agentic music production environment built on Claude Code"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "pretty-midi>=0.2.10",
    "music21>=9.1",
    "librosa>=0.10",
    "pydantic>=2.5",
    "pyloudnorm>=0.1",
    "structlog>=24.1",
    "pyyaml>=6.0",
    "numpy>=1.24",
    "scipy>=1.11",
    "soundfile>=0.12",
    "click>=8.1",
    "pedalboard>=0.9",
    "sounddevice>=0.4.6",
    "watchdog>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "mypy>=1.8",
    "ruff>=0.2",
    "types-PyYAML",
    "pre-commit>=3.6",
]
neural = [
    "torch>=2.0",
    "transformers>=4.40",
]
live = [
    "mido>=1.3",
    "python-rtmidi>=1.5",
]
annotate = [
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "jinja2>=3.1",
]

[project.scripts]
yao = "cli.main:cli"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.mypy]
strict = true
python_version = "3.11"
packages = ["yao"]
mypy_path = "src"

[[tool.mypy.overrides]]
module = [
    "pretty_midi",
    "pretty_midi.*",
    "pydantic",
    "pydantic.*",
    "structlog",
    "structlog.*",
    "music21",
    "music21.*",
    "librosa",
    "librosa.*",
    "pyloudnorm",
    "pyloudnorm.*",
    "soundfile",
    "soundfile.*",
    "scipy",
    "scipy.*",
    "numpy",
    "numpy.*",
    "yaml",
    "yaml.*",
    "sounddevice",
    "sounddevice.*",
    "watchdog",
    "watchdog.*",
    "pedalboard",
    "pedalboard.*",
    "torch",
    "torch.*",
    "transformers",
    "transformers.*",
    "audiocraft",
    "audiocraft.*",
    "mido",
    "mido.*",
    "rtmidi",
    "rtmidi.*",
    "fastapi",
    "fastapi.*",
    "uvicorn",
    "uvicorn.*",
    "jinja2",
    "jinja2.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "yao.annotate.server"
disallow_untyped_decorators = false

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: integration tests",
    "golden: golden output tests",
    "music_constraints: music theory constraint tests",
    "subagent: subagent evaluation tests",
    "subjective: subjective quality tests (human-rated, skip in CI)",
]
